Fix passwordless logins
In 34db9775 support for logins without authentication was added.
However, in 355ff5ca it got broken by inverting the logic of whether
passwords should be encoded or not, yielding to the encoding of None
as bytes("None"). The result of this conversion then doesn't pass the
`if not password and not allow_agent` check in pythonclient, and as a
result we always try to authenticate (with "None" as a password).